Move existing-credit-card payment method to calypso
