+++ packages/wpcom-checkout/src/index.ts
@@ -18,7 +18,6 @@ export * from './payment-methods/eps';
 export * from './payment-methods/alipay';
 export * from './use-is-web-payment-available';
 export * from './payment-methods/google-pay';
-export * from './payment-methods/existing-credit-card';
 export { isWpComProductRenewal } from './is-wpcom-product-renewal';
 export { isValueTruthy } from './is-value-truthy';
 export * from './checkout-labels';
